refactor(ai): migrate to multi-provider LLM support with Genkit
- Replace direct Gemini SDK with Firebase Genkit for unified LLM integration
- Add support for Google AI, OpenAI, Anthropic, and Ollama providers
- Remove circuit breaker from agent executor, delegate to MEB store
- Add telemetry package and integrate logging sink in store manager
- Refactor query execution with single/multi-atom query handling and LFTJ/sequential join fallbacks
- Enhance ingestion with project-scoped topic IDs and batch hydration
- Update server to use AIService instead of GeminiService
- Add semantic search filtering and improved graph pathfinding
- Remove HTML sanitization from validation to avoid corrupting Datalog queries
- Update dependencies and environment variables for LLM configuration

BREAKING CHANGE: Environment variables changed from GEMINI_API_KEY/MODEL to LLM_API_KEY/MODEL/PROVIDER, and NewServer no longer takes apiKey parameter.

+++ pkg/ingest/extractor.go
@@ -277,8 +277,8 @@ func (e *TreeSitterExtractor) processSymbols(bundle *AnalysisBundle, symbols []S
 			Content: []byte(sym.Content),
 			Metadata: map[string]any{
 				"file":       relPath,
-				"start_line": sym.StartLine,
-				"end_line":   sym.EndLine,
+				"start_line": int32(sym.StartLine),
+				"end_line":   int32(sym.EndLine),
 				"package":    filePackage,
 				"tags":       tags,
 			},
